Add support for HTTP 415 and 422
RFC 7231 HTTP 415, RFC 4918 HTTP 422 are both suited for use with POST, PUT, PATCH when the payload is off in some way. See e.g. https://datatracker.ietf.org/doc/html/rfc5789#section-2.2. But likely HTTP 415 should actually be supported by the router code when the path is correct, but the Content-Type is not supported.