feat(errors): add errors for HTTP 415 and 422
Two new errors are exported by the errors module:

- $errors:UNSUPPORTED_MEDIA_TYPE (mapping to status code 415)
- $errors:UNPROCESSABLE_ENTITY (mapping to status code 422)

RFC 7231 HTTP 415, RFC 4918 HTTP 422 are both suited for use with POST, PUT
and PATCH when the payload is off in some way.
See e.g. [RFC5789](]https://datatracker.ietf.org/doc/html/rfc5789#section-2.2).

But likely HTTP 415 should actually be supported by the router code when the
path is correct, but the Content-Type is not supported.