Implement more robust support for mediaType detection e.g. to support RFC 6902
When you submit data with content-type application/json-patch+json, the router does not handle the contents as json but just sends it in as string. There are more mediaTypes like this, e.g. application/fhir+json and application/fhir+xml. This updates tries to be as graceful as possible with mediaTypes that declare +json or +xml.