feat(router): improve media type handling
closes #37

More robust handling of  media-types.

The request map is enhanced with following properties

 - media-type: normalized Content-Type header value
- charset: if it was set in the Content-Type header,
- registry: "application", "image", "text", ...
- format: the format of the body
  ("json", "xml", "form-data", "binary", "zip", ...

The above properties are only available if the matched route expects a body with the request.

This information is also used to determine how to handle body contents.

Improve function router:method-for-content-type to select
the appropriat output serializattion for
more media types.

Thanks to Alexander Henket to bring this to my attention.