fix(parameters): fail if multiple values are sent to non-array
fixes #86

A request like `/my/api/endpoint?a=1&a=2` will now fail unless the query parameter `a` is declared an array.
This is done to support both API implementors and users to find problems easier which would otherwise go undetected. The first value of a would be used and the second one just discarded silently.